training_tool/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
PyQt5>=5.15.0
Pillow>=9.0.0
numpy>=1.24.0
opencv-python>=4.8.0
matplotlib>=3.7.0
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.31.0
flask>=2.3.0
flask-cors>=4.0.0
